Types in Ushort folders moved to appropriate folders.
